# docs/collection/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Sistema de Autoatendimento da Lanchonete",
        "description": "Documentacao automatica via Swagger e Redoc",
        "version": "1.0.0"
    },
    "paths": {
...
        "/clientes/": {
            "get": {
                "tags": ["clientes"],
                "summary": "Listar Clientes",
                "operationId": "listar_clientes_clientes__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ClienteResponseSchema"
                                    },
                                    "type": "array",
                                    "title": "Response Listar Clientes Clientes  Get"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": ""
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["clientes"],
                "summary": "Criar Cliente",
                "operationId": "criar_cliente_clientes__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteCreateSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponseSchema"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Erro de integridade ao criar cliente"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/clientes/cpf/{cpf}": {
            "get": {
                "tags": ["clientes"],
                "summary": "Buscar Cliente Por Cpf",
                "operationId": "buscar_cliente_por_cpf_clientes_cpf__cpf__get",
                "parameters": [{
                    "name": "cpf",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "title": "Cpf"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponseSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Cliente não encontrado"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": ""
                                }
                            }
                        }
                    }
                }
            }
        },
        "/clientes/{cliente_id}": {
            "get": {
                "tags": ["clientes"],
                "summary": "Buscar Cliente",
                "operationId": "buscar_cliente_clientes__cliente_id__get",
                "parameters": [{
                    "name": "cliente_id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "title": "Cliente Id"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponseSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Cliente não encontrado"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": ""
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["clientes"],
                "summary": "Atualizar Cliente",
                "operationId": "atualizar_cliente_clientes__cliente_id__put",
                "parameters": [{
                    "name": "cliente_id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "title": "Cliente Id"
                    }
                }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteUpdateSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClienteResponseSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Cliente não encontrado"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Erro de integridade ao atualizar o cliente"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["clientes"],
                "summary": "Deletar Cliente",
                "operationId": "deletar_cliente_clientes__cliente_id__delete",
                "parameters": [{
                    "name": "cliente_id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "integer",
                        "title": "Cliente Id"
                    }
                }],
                "responses": {
                    "204": {
                        "description": "Pedido deletado com sucesso",
                        "content": {
                            "application/json": {
                                "example": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Cliente não encontrado"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "example": {
                                    "message": "Erro de integridade ao deletar o cliente"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClienteCreateSchema": {
                "properties": {
                    "nome": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 3,
                        "title": "Nome"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "telefone": {
                        "anyOf": [{
                            "type": "string",
                            "maxLength": 11,
                            "minLength": 10,
                            "pattern": "^\\d{10,11}$"
                        }, {
                            "type": "null"
                        }],
                        "title": "Telefone"
                    },
                    "cpf": {
                        "type": "string",
                        "maxLength": 11,
                        "minLength": 11,
                        "pattern": "^\\d{11}$",
                        "title": "Cpf"
                    }
                },
                "type": "object",
                "required": ["nome", "email", "cpf"],
                "title": "ClienteCreateSchema"
            },
            "ClienteResponseSchema": {
                "properties": {
                    "cliente_id": {
                        "type": "integer",
                        "title": "Cliente Id"
                    },
                    "nome": {
                        "anyOf": [{
                            "type": "string"
                        }, {
                            "type": "null"
                        }],
                        "title": "Nome"
                    },
                    "email": {
                        "anyOf": [{
                            "type": "string"
                        }, {
                            "type": "null"
                        }],
                        "title": "Email"
                    },
                    "telefone": {
                        "anyOf": [{
                            "type": "string"
                        }, {
                            "type": "null"
                        }],
                        "title": "Telefone"
                    },
                    "cpf": {
                        "anyOf": [{
                            "type": "string"
                        }, {
                            "type": "null"
                        }],
                        "title": "Cpf"
                    }
                },
                "type": "object",
                "required": ["cliente_id", "nome", "email", "telefone", "cpf"],
                "title": "ClienteResponseSchema"
            },
            "ClienteUpdateSchema": {
                "properties": {
                    "nome": {
                        "anyOf": [{
                            "type": "string",
                            "maxLength": 100,
                            "minLength": 3
                        }, {
                            "type": "null"
                        }],
                        "title": "Nome"
                    },
                    "email": {
                        "anyOf": [{
                            "type": "string",
                            "format": "email"
                        }, {
                            "type": "null"
                        }],
                        "title": "Email"
                    },
                    "telefone": {
                        "anyOf": [{
                            "type": "string",
                            "maxLength": 15,
                            "minLength": 10
                        }, {
                            "type": "null"
                        }],
                        "title": "Telefone"
                    },
                    "cpf": {
                        "anyOf": [{
                            "type": "string",
                            "maxLength": 11,
                            "minLength": 11,
                            "pattern": "^\\d{11}$"
                        }, {
                            "type": "null"
                        }],
                        "title": "Cpf"
                    }
                },
                "type": "object",
                "title": "ClienteUpdateSchema"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [{
                                "type": "string"
                            }, {
                                "type": "integer"
                            }]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": ["loc", "msg", "type"],
                "title": "ValidationError"
            }
        }
    }
}